chore: genericify operator provider, consolidate env, handle async correctly
